refactor: :fire: change Modal to Dialog
