+++ src/components/Dialog/apis/apiEvents.vue
@@ -1,0 +1,19 @@
+<template>
+	<ApiEventTable :events="data" />
+</template>
+
+<script setup>
+import ApiEventTable from '@/components/api-table/apiEvent.vue'
+const data = [
+	{
+		name: 'close',
+		description: 'Close the modal',
+		parameters: '',
+	},
+	{
+		name: 'cancel',
+		description: 'Do a show on modal',
+		parameters: '',
+	},
+]
+</script>
