+++ src/components/Dialog/apis/contentApiProperties.vue
