+++ src/components/Dialog/longContent.vue
