--- src/content/docs/components/modal.mdx
@@ -1,54 +1,0 @@
----
-title: Dialog
----
-
-The `Dialog` component is a modal interface element used to capture user attention and present important content, actions, or forms in a focused overlay. It is typically used for confirmations, alerts, input prompts, or custom interactive content without navigating away from the current page.
-
-import { Code } from '@astrojs/starlight/components'
-import { Tabs, TabItem } from '@astrojs/starlight/components'
-import ApiProperties from '@/components/Modal/apis/apiProperties.vue'
-import ContentApiProperties from '@/components/Modal/apis/contentApiProperties.vue'
-
-import SBasic from '@/components/Modal/basic.vue'
-import SBasicRaw from '@/components/Modal/basic.vue?raw'
-
-import SContent from '@/components/Modal/longContent.vue'
-import SContentRaw from '@/components/Modal/longContent.vue?raw'
-
-import SAction from '@/components/Modal/multiAction.vue'
-import SActionRaw from '@/components/Modal/multiAction.vue?raw'
-
-## Component Usage
-
-### Basic
-
-<SBasic client:only="vue" class="mt-4" />
-<Code code={SBasicRaw} lang="vue" title="basic.vue" />
-
-### Long Content
-
-<SContent client:only="vue" class="mt-4" />
-<Code code={SContentRaw} lang="vue" title="longContent.vue" />
-
-### Multi Action
-
-<SAction client:only="vue" class="mt-4" />
-<Code code={SActionRaw} lang="vue" title="multiAction.vue" />
-
-## API
-
-### Dialog
-
-<Tabs>
-	<TabItem label="Properties">
-		<ApiProperties client:only="vue" />
-	</TabItem>
-</Tabs>
-
-### Dialog Content
-
-<Tabs>
-	<TabItem label="Properties">
-		<ContentApiProperties client:only="vue" />
-	</TabItem>
-</Tabs>
